proposed fix for cardboard button when 'poll=true'
See: issue #4 

methodology: we still primarily rely on polling when poll=true, but we also add the `selectstart` and `selectend` listeners in and specifically listen for the `gaze` controller; when we detect those movements, we update a fake psuedo-controller record. `syntheticGamepad()` recognizes when it is getting an update from the cardboard-gaze controller, and instead of just pushing two hardcoded button values, it refers to this psuedo-controller record and passes that along--in this way, the psuedo-controller record is picked up within the polling loop like other controllers in this method.